+++ tests/ops/models/privacy_request/test_privacy_request.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timedelta, timezone
 from time import sleep
-from typing import Any, Dict, List
+from typing import Any, Dict, List, Tuple
 from uuid import uuid4
 
 import pytest
@@ -20,9 +20,10 @@ from fides.api.models.privacy_request import (
     PrivacyRequest,
     PrivacyRequestError,
     PrivacyRequestNotifications,
+    ProvidedIdentity,
     can_run_checkpoint,
 )
-from fides.api.schemas.policy import ActionType, CurrentStep
+from fides.api.schemas.policy import CurrentStep
 from fides.api.schemas.privacy_request import (
     CheckpointActionRequired,
     CustomPrivacyRequestField,
@@ -40,6 +41,31 @@ from fides.api.util.constants import API_DATE_FORMAT
 from fides.config import CONFIG
 
 paused_location = CollectionAddress("test_dataset", "test_collection")
+
+
+def test_provided_identity_to_identity(
+    provided_identity_and_consent_request: Tuple,
+) -> None:
+    provided_identity = provided_identity_and_consent_request[0]
+    identity = provided_identity.as_identity_schema()
+    assert identity.email == "test@email.com"
+
+
+def test_blank_provided_identity_to_identity(
+    empty_provided_identity: ProvidedIdentity,
+) -> None:
+    identity = empty_provided_identity.as_identity_schema()
+    assert identity.email is None
+
+
+def test_custom_provided_identity_to_identity(
+    custom_provided_identity: ProvidedIdentity,
+) -> None:
+    identity = custom_provided_identity.as_identity_schema()
+    assert identity.customer_id == LabeledIdentity(
+        label=custom_provided_identity.field_label,
+        value=custom_provided_identity.encrypted_value.get("value"),
+    )
 
 
 def test_privacy_request(
@@ -1137,19 +1163,3 @@ class TestPrivacyRequestCustomIdentities:
             customer_id=LabeledIdentity(label="Custom ID", value=123),
             account_id=LabeledIdentity(label="Account ID", value="456"),
         )
-
-
-class TestGetCeleryTaskRequestTaskIds:
-    def test_get_celery_task_request_task_ids(self, privacy_request, request_task):
-        """Not all request tasks have celery task ids in this test -"""
-
-        assert privacy_request.get_request_task_celery_task_ids() == []
-
-        cache_task_tracking_key(request_task.id, "test_celery_task_key")
-        root_task = privacy_request.get_root_task_by_action(ActionType.access)
-        cache_task_tracking_key(root_task.id, "test_root_task_celery_key")
-
-        assert set(privacy_request.get_request_task_celery_task_ids()) == {
-            "test_celery_task_key",
-            "test_root_task_celery_key",
-        }
